Re-construct ast::de module
