+++ ruststep/src/ast/de/parameter.rs
@@ -6,23 +6,6 @@ use serde::{
     forward_to_deserialize_any,
 };
 
-impl<'de, 'record> de::Deserializer<'de> for &'record Record {
-    type Error = crate::error::Error;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::Visitor<'de>,
-    {
-        visitor.visit_map(self.into_deserializer())
-    }
-
-    forward_to_deserialize_any! {
-        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
-        bytes byte_buf option unit unit_struct newtype_struct seq tuple
-        struct tuple_struct map enum identifier ignored_any
-    }
-}
-
 impl<'de, 'param> de::Deserializer<'de> for &'param Parameter {
     type Error = crate::error::Error;
 
@@ -50,5 +33,38 @@ impl<'de, 'param> de::Deserializer<'de> for &'param Parameter {
         bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
         bytes byte_buf option unit unit_struct newtype_struct seq tuple
         struct tuple_struct map enum identifier ignored_any
+    }
+}
+
+#[derive(Debug)]
+pub struct SeqDeserializer {
+    parameters: Vec<Parameter>,
+}
+
+impl SeqDeserializer {
+    pub fn new(parameters: &[Parameter]) -> Self {
+        SeqDeserializer {
+            parameters: parameters.iter().rev().cloned().collect(),
+        }
+    }
+}
+
+impl<'de> de::SeqAccess<'de> for SeqDeserializer {
+    type Error = crate::error::Error;
+
+    fn size_hint(&self) -> Option<usize> {
+        Some(self.parameters.len())
+    }
+
+    fn next_element_seed<T>(&mut self, seed: T) -> Result<Option<T::Value>, Self::Error>
+    where
+        T: de::DeserializeSeed<'de>,
+    {
+        if let Some(last) = self.parameters.pop() {
+            let value = seed.deserialize(&last)?;
+            Ok(Some(value))
+        } else {
+            Ok(None)
+        }
     }
 }
