+++ ruststep/src/ast/de/record.rs
@@ -1,0 +1,91 @@
+use crate::ast::*;
+use serde::{
+    de::{self, IntoDeserializer},
+    forward_to_deserialize_any,
+};
+
+impl<'de, 'record> de::Deserializer<'de> for &'record Record {
+    type Error = crate::error::Error;
+
+    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
+    where
+        V: de::Visitor<'de>,
+    {
+        visitor.visit_map(self.into_deserializer())
+    }
+
+    forward_to_deserialize_any! {
+        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
+        bytes byte_buf option unit unit_struct newtype_struct seq tuple
+        struct tuple_struct map enum identifier ignored_any
+    }
+}
+
+/// Deserializer corresponding to a single-key map like `{ "A": [1.0, 2.0] }`
+#[derive(Debug)]
+pub struct RecordDeserializer {
+    key: Option<String>,
+    value: Parameter,
+}
+
+impl<'de, 'record> de::IntoDeserializer<'de, crate::error::Error> for &'record Record {
+    type Deserializer = RecordDeserializer;
+    fn into_deserializer(self) -> RecordDeserializer {
+        RecordDeserializer {
+            key: Some(self.name.to_string()),
+            value: self.parameter.clone(),
+        }
+    }
+}
+
+impl<'de> de::Deserializer<'de> for RecordDeserializer {
+    type Error = crate::error::Error;
+
+    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
+    where
+        V: de::Visitor<'de>,
+    {
+        visitor.visit_map(self)
+    }
+
+    forward_to_deserialize_any! {
+        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
+        bytes byte_buf option unit unit_struct newtype_struct seq tuple
+        struct tuple_struct map enum identifier ignored_any
+    }
+}
+
+impl RecordDeserializer {
+    pub fn new(key: &str, value: Parameter) -> Self {
+        RecordDeserializer {
+            key: Some(key.to_string()),
+            value,
+        }
+    }
+}
+
+// Entry point of `visit_map`
+impl<'de> de::MapAccess<'de> for RecordDeserializer {
+    type Error = crate::error::Error;
+
+    fn next_key_seed<K>(&mut self, seed: K) -> Result<Option<K::Value>, Self::Error>
+    where
+        K: de::DeserializeSeed<'de>,
+    {
+        if let Some(key) = self.key.take() {
+            let key: de::value::StrDeserializer<Self::Error> = key.as_str().into_deserializer();
+            let key: K::Value = seed.deserialize(key)?;
+            Ok(Some(key))
+        } else {
+            Ok(None)
+        }
+    }
+
+    fn next_value_seed<V>(&mut self, seed: V) -> Result<V::Value, Self::Error>
+    where
+        V: de::DeserializeSeed<'de>,
+    {
+        let value: V::Value = seed.deserialize(&self.value)?;
+        Ok(value)
+    }
+}
